+++ modules/registration_module/registration_functions.py
@@ -1242,7 +1242,7 @@ def manual_edit_main(image_path, matrix_path):
 
     viewer = napari.Viewer()
     # assuming a FTCZYX image:
-    viewer.add_image(image.image, name="image", channel_axis=2)
+    viewer.add_image(image.image, channel_axis=2, name=['Image [' + x + ']' for x in image.channel_names] if image.channel_names else 'Image')
     # channel axis is already used as channel_axis (layers) => it is not in viewer.dims:
     viewer.dims.axis_labels = ('F', 'T', 'Z', 'Y', 'X')